docs(odin-html) - more text styles md file translation (#745)
# Description
Turkish translation of "more-text-styles" md file

Note: I didn't change some of the headers in the md file because they
are related to the specific tag names. If it's necessary, drop a message
about it.

### Checklist

Turkish transation of "more-text-styles" md file

Note: I didn't change some of the headers in the md file because they
are related to the specific tag names. If it's necessary, drop a message
about it.

- [ x] discord username: `doguhan_i#0001`
- [ ] Closes #<issue-number>
- [ ] PR must be created for an issue from issues under "In progress"
column from [our project
board](https://github.com/orgs/kamp-us/projects/2/views/1).
- [ x] A descriptive and understandable title: The PR title should
clearly describe the nature and purpose of the changes. The PR title
should be the first thing displayed when the PR is opened. And it should
follow the semantic commit rules, and should include the
app/package/service name in the title. For example, a title like
"docs(@kampus-apps/pano): Add README.md" can be used.
- [ ] Related file selection: Only relevant files should be touched and
no other files should be affected.
- [ ] I ran `npx turbo run` at the root of the repository, and build was
successful.
- [ ] I installed the npm packages using `npm install --save-exact
<package>` so my package is pinned to a specific npm version. Leave
empty if no package was installed. Leave empty if no package was
installed with this PR.

### How were these changes tested?

Please describe the tests you did to test the changes you made. Please
also specify your test configuration.

---------

Co-authored-by: Huseyin Gulec <itshuseyingulec@gmail.com>